tickets/src/events/listeners/order_created_listener.ts
import { Listener, OrderCreatedEvent, Subjects } from "@richardcap/common";
import { queueGroupName } from "./queue_group_name";
import { Message } from "node-nats-streaming";
import { Ticket } from "../../models/ticket";
import { TicketUpdatedPublisher } from "../publishers/ticket_updated_publisher";

export class OrderCreatedListener extends Listener<OrderCreatedEvent> {
    subject: Subjects.OrderCreated = Subjects.OrderCreated;
    queueGroupName = queueGroupName;

    async onMessage(data: OrderCreatedEvent['data'], msg: Message) {
        // Find the ticket
        const ticket = await Ticket.findById(data.ticket.id);
        if (!ticket) {
            throw new Error('Ticket not found');
        }
        // Mark the ticket as being reserved
        ticket.set({
            orderId: data.id
        });
        // Save the ticket
        await ticket.save();
        // Publish an event
        await new TicketUpdatedPublisher(this.client).publish({
            id: ticket.id,
            title: ticket.title,
            price: ticket.price,
            userId: ticket.userId,
            orderId: ticket.orderId,
            version: ticket.version
        });

        // Ack the msg
        msg.ack();
    }
}